% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_read.R
\name{kfa_read_pointcloud}
\alias{kfa_read_pointcloud}
\title{Read KyFromAbove Point Cloud for an Area of Interest}
\usage{
kfa_read_pointcloud(aoi, phase = 2L)
}
\arguments{
\item{aoi}{An \link[sf:sf]{sf::sf} object or numeric bbox.}

\item{phase}{Integer. 1, 2, or 3.}
}
\value{
A \link[lidR:LAS-class]{lidR::LAS} object.
}
\description{
Finds point cloud tiles (LAZ for Phase 1, COPC for Phase 2/3)
covering the AOI and reads them via \code{\link[lidR:readLAS]{lidR::readLAS()}}.
}
\examples{
\dontshow{if (aboveR::has_s3_access()) withAutoprint(\{ # examplesIf}
\donttest{
las <- kfa_read_pointcloud(
  aoi = c(-84.55, 37.95, -84.54, 37.96),
  phase = 2
)
}
\dontshow{\}) # examplesIf}
}
