% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclamation.R
\name{surface_roughness}
\alias{surface_roughness}
\title{Compute Surface Roughness of a DEM}
\usage{
surface_roughness(dem, window = 5L)
}
\arguments{
\item{dem}{A \link[terra:SpatRaster-class]{terra::SpatRaster}.}

\item{window}{Integer. Size of the moving window (number of cells).
Must be odd. Default \code{5}.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} of local roughness values
(standard deviation of elevation).
}
\description{
Calculates surface roughness as the standard deviation of elevation
within a moving window. Rougher surfaces indicate unreclaimed terrain,
active construction, or natural heterogeneity.
}
\examples{
dem <- terra::rast(system.file("extdata/dem_before.tif", package = "aboveR"))
rough <- surface_roughness(dem, window = 5)
terra::plot(rough)
}
