% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboveR-package.R
\docType{package}
\name{aboveR-package}
\alias{aboveR}
\alias{aboveR-package}
\title{aboveR: LiDAR Terrain Analysis and Change Detection from Above}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Terrain change detection, cut and fill volume estimation, terrain profiling, reclamation monitoring, erosion analysis, and flood risk assessment from LiDAR point clouds and DEMs. Built on 'lidR' for point cloud I/O and 'terra' for raster operations. Includes access utilities for 'KyFromAbove' cloud-native elevation data on AWS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/null-island-labs/aboveR}
  \item Report bugs at \url{https://github.com/null-island-labs/aboveR/issues}
}

}
\author{
\strong{Maintainer}: Chris Lyons \email{chris@nullislandlabs.dev}

}
\keyword{internal}
