% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{flood_colors}
\alias{flood_colors}
\title{Flood Depth Color Ramp}
\usage{
flood_colors(n = 100L)
}
\arguments{
\item{n}{Integer. Number of colors. Default \code{100}.}
}
\value{
Character vector of hex color values.
}
\description{
Light-to-dark blue color ramp for flood depth visualisation.
}
\examples{
cols <- flood_colors(10)
image(matrix(1:10), col = cols)
}
