% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_counties.R
\name{kfa_list_counties}
\alias{kfa_list_counties}
\title{List Available Kentucky Counties}
\usage{
kfa_list_counties()
}
\value{
Character vector of county names sorted alphabetically.
}
\description{
Returns a character vector of all 120 Kentucky county names.
}
\examples{
counties <- kfa_list_counties()
head(counties)
}
