% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfa_stac.R
\name{kfa_stac_search}
\alias{kfa_stac_search}
\title{Search KyFromAbove STAC Catalog}
\usage{
kfa_stac_search(aoi, collection = NULL, datetime = NULL)
}
\arguments{
\item{aoi}{An \link[sf:sf]{sf::sf} object or numeric bbox.}

\item{collection}{Character. STAC collection ID.}

\item{datetime}{Character. ISO 8601 datetime or range.}
}
\value{
A tibble of STAC items with asset URLs.
}
\description{
Queries the KyFromAbove STAC catalog for items matching an area
of interest and product type. Requires the 'rstac' package.
}
